Migrate worker to typescript.

+++ worker/consumer/transcodeMP3.ts
@@ -1,4 +1,6 @@
 import { streamFromBucket, streamToBucket } from "gridfire-worker/controllers/storage.js";
+import { ReleaseContext } from "gridfire-worker/types/index.js";
+import assert from "assert/strict";
 import { ffmpegEncodeMP3FromStream } from "gridfire-worker/consumer/ffmpeg.js";
 import fs from "fs";
 import mongoose from "mongoose";
@@ -10,7 +12,11 @@ const { BUCKET_FLAC, BUCKET_MP3, TEMP_PATH } = process.env;
 const Release = mongoose.model("Release");
 const fsPromises = fs.promises;
 
-const transcodeMP3 = async ({ releaseId, trackId, userId }) => {
+assert(BUCKET_FLAC, "BUCKET_FLAC env var missing.");
+assert(BUCKET_MP3, "BUCKET_MP3 env var missing.");
+assert(TEMP_PATH, "TEMP_PATH env var missing.");
+
+const transcodeMP3 = async ({ releaseId, trackId, userId }: ReleaseContext) => {
   let mp3FilePath;
 
   try {
@@ -34,7 +40,10 @@ const transcodeMP3 = async ({ releaseId, trackId, userId }) => {
     postMessage({ type: "pipelineError", stage: "mp3", trackId, userId });
   } finally {
     console.log("Removing temp MP3 stage file:", mp3FilePath);
-    await fsPromises.unlink(mp3FilePath);
+
+    if (mp3FilePath) {
+      await fsPromises.unlink(mp3FilePath);
+    }
   }
 };
